move cat/bcat unit tests
